requirements.txt
alembic==1.8.1
bcrypt==4.0.1
blinker==1.5
cffi==1.15.1
click==8.1.3
cryptography==38.0.4
Deprecated==1.2.13
dnspython==2.3.0
email-validator==1.3.0
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.1
httpagentparser==1.9.5
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.4.2
PyJWT==2.6.0
Mako==1.2.4
MarkupSafe==2.1.1
pandas==1.5.3
python-dotenv==0.21.0
SQLAlchemy==1.4.44
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==3.0.1
